# swagger/user.routes.yaml
openapi: '3.0.0'
info:
  description: 'API documentation for user operations'
  version: '1.0.0'
  title: 'User API'
servers:
  - url: 'http://localhost:3000/api'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but good practice
...
paths:
...
  /random-joke:
    get:
      tags:
        - 'User'
      summary: 'Random joke'
      description: 'Get a random joke'
      operationId: 'randomJoke'
      security:
        - bearerAuth: []
      responses:
        200:
          description: 'Random joke'
        default:
          description: 'Unexpected error'